requirements.txt
# Core Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0

# LLM Providers
openai>=1.0.0
groq>=0.9.0

# Vector Database & Retrieval
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
numpy>=1.26.0
scipy>=1.12.0

# Text Processing & Document Parsing
pypdf>=5.0.0
unstructured>=0.16.0
python-magic>=0.4.27
markdown>=3.7

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Web UI
streamlit>=1.40.0
streamlit-shadcn-ui>=0.1.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tenacity>=9.0.0
aiohttp>=3.9.0
tqdm>=4.66.0
python-json-logger>=2.0.7

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.6.0

# Type Checking
mypy>=1.8.0
types-requests>=2.31.0

# Security
bandit>=1.7.0

# Optional: For GPU acceleration (uncomment if using faiss-gpu)
# faiss-gpu>=1.8.0